Lab_12/backend/requirements.txt
altgraph
asgiref==3.6.0
astroid==2.12.13
attrs==22.2.0
dill==0.3.6
Django==4.1.5
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.4.1
flake8==5.0.4
future
importlib-metadata==6.0.0
inflection==0.5.1
isort==5.10.1
jsonschema==4.17.3
lazy-object-proxy==1.8.0
macholib
Markdown==3.4.1
mccabe==0.7.0
MouseInfo==0.1.3
platformdirs==2.5.4
pycodestyle==2.9.1
pyflakes==2.5.0
PyGetWindow==0.0.9
PyJWT==2.6.0
pylint==2.15.7
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
pytweening==1.0.4
pytz==2022.7.1
PyYAML==6.0
rubicon-objc==0.4.3
six
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
uritemplate==4.1.1
wrapt==1.14.1
zipp==3.13.0
gunicorn==20.1.0